Chord List, ChordApi, ChordService, Component Tree
The chord list was modified to be more general. It can now be used to get user specific and public chord lists.

The chord api was fixed to accomate serving lists of public cords and user specific chords.

The public component hiarachy was fleshed out. The login function was renamed.